+++ casanovo/utils.py
@@ -4,11 +4,20 @@ import logging
 import os
 import platform
 import re
-from typing import Tuple
+import socket
+import sys
+from datetime import datetime
+from typing import Tuple, Dict, List, Optional
 
+import numpy as np
+import pandas as pd
 import psutil
 import torch
 
+from .data.ms_io import PepSpecMatch
+
+
+SCORE_BINS = [0.0, 0.5, 0.9, 0.95, 0.99]
 
 logger = logging.getLogger("casanovo")
 
@@ -66,3 +75,179 @@ def split_version(version: str) -> Tuple[str, str, str]:
     """
     version_regex = re.compile(r"(\d+)\.(\d+)\.*(\d*)(?:.dev\d+.+)?")
     return tuple(g for g in version_regex.match(version).groups())
+
+
+def get_score_bins(
+    scores: pd.Series, score_bins: List[float]
+) -> Dict[float, int]:
+    """
+    Get binned confidence scores
+
+    From a list of confidence scores, return a dictionary mapping each
+    confidence score to the number of spectra with a confidence greater
+    than or equal to it.
+
+    Parameters
+    ----------
+    scores: pd.Series
+        Series of assigned peptide scores.
+    score_bins: List[float]
+        Confidence scores to map.
+
+    Returns
+    -------
+    score_bin_dict: Dict[float, int]
+        Dictionary mapping each confidence score to the number of spectra
+        with a confidence greater than or equal to it.
+    """
+    return {score: (scores >= score).sum() for score in score_bins}
+
+
+def get_peptide_lengths(sequences: pd.Series) -> np.ndarray:
+    """
+    Get a numpy array containing the length of each peptide sequence
+
+    Parameters
+    ----------
+    sequences: pd.Series
+        Series of peptide sequences.
+
+    Returns
+    -------
+    sequence_lengths: np.ndarray
+        Numpy array containing the length of each sequence, listed in the
+        same order that the sequences are provided in.
+    """
+    # Mass modifications do not contribute to sequence length
+    # FIXME: If PTMs are represented in ProForma notation this filtering
+    #  operation needs to be reimplemented
+    return sequences.str.replace(r"[^a-zA-Z]", "", regex=True).apply(len)
+
+
+def get_report_dict(
+    results_table: pd.DataFrame, score_bins: List[float] = SCORE_BINS
+) -> Optional[Dict]:
+    """
+    Generate sequencing run report
+
+    Parameters
+    ----------
+    results_table: pd.DataFrame
+        Parsed spectrum match table
+    score_bins: List[float], Optional
+        Confidence scores for creating confidence CMF, see get_score_bins
+
+    Returns
+    -------
+    report_gen: Dict
+        Generated report represented as a dictionary, or None if no
+        sequencing predictions were logged
+    """
+    if results_table.empty:
+        return None
+
+    peptide_lengths = get_peptide_lengths(results_table["sequence"])
+    min_length, med_length, max_length = np.quantile(
+        peptide_lengths, [0, 0.5, 1]
+    )
+    return {
+        "num_spectra": len(results_table),
+        "score_bins": get_score_bins(results_table["score"], score_bins),
+        "max_sequence_length": max_length,
+        "min_sequence_length": min_length,
+        "median_sequence_length": med_length,
+    }
+
+
+def log_run_report(
+    start_time: Optional[int] = None, end_time: Optional[int] = None
+) -> None:
+    """
+    Log general run report
+
+    Parameters
+    ----------
+    start_time : Optional[int], default=None
+        The start time of the sequencing run in seconds since the epoch.
+    end_time : Optional[int], default=None
+        The end time of the sequencing run in seconds since the epoch.
+    """
+    logger.info("======= End of Run Report =======")
+    if start_time is not None and end_time is not None:
+        start_datetime = datetime.fromtimestamp(start_time)
+        end_datetime = datetime.fromtimestamp(end_time)
+        delta_datetime = end_datetime - start_datetime
+        logger.info(
+            "Run Start Time: %s",
+            start_datetime.strftime("%y/%m/%d %H:%M:%S"),
+        )
+        logger.info(
+            "Run End Time: %s", end_datetime.strftime("%y/%m/%d %H:%M:%S")
+        )
+        logger.info("Time Elapsed: %s", delta_datetime)
+
+    logger.info("Executed Command: %s", " ".join(sys.argv))
+    logger.info("Executed on Host Machine: %s", socket.gethostname())
+
+    if torch.cuda.is_available():
+        gpu_util = torch.cuda.max_memory_allocated()
+        logger.info("Max GPU Memory Utilization: %d MiB", gpu_util >> 20)
+
+
+def log_sequencing_report(
+    predictions: List[PepSpecMatch],
+    start_time: Optional[int] = None,
+    end_time: Optional[int] = None,
+    score_bins: List[float] = SCORE_BINS,
+) -> None:
+    """
+    Log sequencing run report
+
+    next_prediction : Tuple[
+        str, Tuple[str, str], float, float, float, float, str
+    ]
+        PSM predictions
+    start_time : Optional[int], default=None
+        The start time of the sequencing run in seconds since the epoch.
+    end_time : Optional[int], default=None
+        The end time of the sequencing run in seconds since the epoch.
+    score_bins: List[float], Optional
+        Confidence scores for creating confidence score distribution,
+        see get_score_bins
+    """
+    log_run_report(start_time=start_time, end_time=end_time)
+    run_report = get_report_dict(
+        pd.DataFrame(
+            {
+                "sequence": [psm.sequence for psm in predictions],
+                "score": [psm.peptide_score for psm in predictions],
+            }
+        ),
+        score_bins=score_bins,
+    )
+
+    if run_report is None:
+        logger.warning(
+            "No predictions were logged, this may be due to an error"
+        )
+    else:
+        num_spectra = run_report["num_spectra"]
+        logger.info("Sequenced %s spectra", num_spectra)
+        logger.info("Score Distribution:")
+        for score, pop in sorted(run_report["score_bins"].items()):
+            logger.info(
+                "%s spectra (%.2f%%) scored ≥ %.2f",
+                pop,
+                pop / num_spectra * 100,
+                score,
+            )
+
+        logger.info(
+            "Min Peptide Length: %d", run_report["min_sequence_length"]
+        )
+        logger.info(
+            "Max Peptide Length: %d", run_report["max_sequence_length"]
+        )
+        logger.info(
+            "Median Peptide Length: %d", run_report["median_sequence_length"]
+        )