+++ tests/unit_tests/test_runner.py
@@ -1,5 +1,7 @@
 """Unit tests specifically for the model_runner module."""
 
+from pathlib import Path
+
 import pytest
 import torch
 
@@ -10,6 +12,7 @@ from casanovo.denovo.model_runner import ModelRunner
 def test_initialize_model(tmp_path, mgf_small):
     """Test initializing a new or existing model."""
     config = Config()
+    config.model_save_folder_path = tmp_path
     # No model filename given, so train from scratch.
     ModelRunner(config=config).initialize_model(train=True)
 
@@ -53,6 +56,7 @@ def test_save_and_load_weights(tmp_path, mgf_small, tiny_config):
     config.max_epochs = 1
     config.n_layers = 1
     ckpt = tmp_path / "test.ckpt"
+    mztab = tmp_path / "test.mztab"
 
     with ModelRunner(config=config) as runner:
         runner.train([mgf_small], [mgf_small])
@@ -80,7 +84,7 @@ def test_save_and_load_weights(tmp_path, mgf_small, tiny_config):
     with torch.device("meta"):
         with ModelRunner(other_config, model_filename=str(ckpt)) as runner:
             with pytest.raises(NotImplementedError) as err:
-                runner.evaluate([mgf_small])
+                runner.predict([mgf_small], mztab)
 
     assert "meta tensor; no data!" in str(err.value)
 
@@ -92,11 +96,11 @@ def test_save_and_load_weights(tmp_path, mgf_small, tiny_config):
     # Shouldn't work:
     with ModelRunner(other_config, model_filename=str(ckpt)) as runner:
         with pytest.raises(RuntimeError):
-            runner.evaluate([mgf_small])
+            runner.predict([mgf_small], mztab)
 
     # Should work:
     with ModelRunner(config=config, model_filename=str(ckpt)) as runner:
-        runner.evaluate([mgf_small])
+        runner.predict([mgf_small], mztab)
 
 
 def test_save_and_load_weights_deprecated(tmp_path, mgf_small, tiny_config):
@@ -149,3 +153,99 @@ def test_calculate_precision(tmp_path, mgf_small, tiny_config):
 
     assert "valid_aa_precision" in runner.model.history.columns
     assert "valid_pep_precision" in runner.model.history.columns
+
+
+def test_save_final_model(tmp_path, mgf_small, tiny_config):
+    """Test that final model checkpoints are saved."""
+    # Test checkpoint saving when val_check_interval is greater than training steps
+    config = Config(tiny_config)
+    config.val_check_interval = 50
+    model_file = tmp_path / "epoch=19-step=20.ckpt"
+    with ModelRunner(config) as runner:
+        runner.train([mgf_small], [mgf_small])
+
+    assert model_file.exists()
+    Path.unlink(model_file)
+
+    # Test checkpoint saving when val_check_interval is not a factor of training steps
+    config.val_check_interval = 15
+    validation_file = tmp_path / "foobar.best.ckpt"
+    with ModelRunner(config, output_rootname="foobar") as runner:
+        runner.train([mgf_small], [mgf_small])
+
+    assert model_file.exists()
+    assert validation_file.exists()
+
+
+def test_evaluate(
+    tmp_path, mgf_small, mzml_small, mgf_small_unannotated, tiny_config
+):
+    """Test model evaluation during sequencing"""
+    # Train tiny model
+    config = Config(tiny_config)
+    config.max_epochs = 1
+    model_file = tmp_path / "epoch=0-step=1.ckpt"
+    with ModelRunner(config) as runner:
+        runner.train([mgf_small], [mgf_small])
+
+    assert model_file.is_file()
+
+    # Test evaluation with annotated peak file
+    result_file = tmp_path / "result.mztab"
+    with ModelRunner(config, model_filename=str(model_file)) as runner:
+        runner.predict([mgf_small], result_file, evaluate=True)
+
+    assert result_file.is_file()
+    result_file.unlink()
+
+    exception_string = (
+        "Error creating annotated spectrum index. "
+        "This may be the result of having an unannotated MGF file "
+        "present in the validation peak file path list.\n"
+    )
+
+    with pytest.raises(FileNotFoundError):
+        with ModelRunner(config, model_filename=str(model_file)) as runner:
+            runner.predict([mzml_small], result_file, evaluate=True)
+
+    with pytest.raises(TypeError, match=exception_string):
+        with ModelRunner(config, model_filename=str(model_file)) as runner:
+            runner.predict([mgf_small_unannotated], result_file, evaluate=True)
+
+    with pytest.raises(TypeError, match=exception_string):
+        with ModelRunner(config, model_filename=str(model_file)) as runner:
+            runner.predict(
+                [mgf_small_unannotated, mzml_small], result_file, evaluate=True
+            )
+
+    # MzTab with just metadata is written in the case of FileNotFound
+    # or TypeError early exit
+    assert result_file.is_file()
+    result_file.unlink()
+
+    # Test mix of annotated an unannotated peak files
+    with pytest.warns(RuntimeWarning):
+        with ModelRunner(config, model_filename=str(model_file)) as runner:
+            runner.predict([mgf_small, mzml_small], result_file, evaluate=True)
+
+    assert result_file.is_file()
+    result_file.unlink()
+
+    with pytest.raises(TypeError, match=exception_string):
+        with ModelRunner(config, model_filename=str(model_file)) as runner:
+            runner.predict(
+                [mgf_small, mgf_small_unannotated], result_file, evaluate=True
+            )
+
+    assert result_file.is_file()
+    result_file.unlink()
+
+    with pytest.raises(TypeError, match=exception_string):
+        with ModelRunner(config, model_filename=str(model_file)) as runner:
+            runner.predict(
+                [mgf_small, mgf_small_unannotated, mzml_small],
+                result_file,
+                evaluate=True,
+            )
+
+    result_file.unlink()